% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bio_bires.R
\name{bio_bires}
\alias{bio_bires}
\title{Function to plot if results above or below cut-off line}
\usage{
bio_bires(
  x,
  y,
  df.data,
  x.plane,
  y.plane,
  df.plane,
  stepwise = TRUE,
  colours = c("red", "green", "grey"),
  inc.equal = TRUE,
  labels = c("above", "below"),
  type = "point"
)
}
\arguments{
\item{x}{y column name in data}

\item{y}{x column name in data}

\item{df.data}{Data frame containing x and y columns}

\item{x.plane}{column name for the x axis in df.plane}

\item{y.plane}{column name for the y axis in df.plane}

\item{df.plane}{Date frame modelling the plane}

\item{stepwise}{logical whether to plot the cutoff plane as stepwise or smoothed}

\item{colours}{colour vector for higher, lower and plane values (default=c("green", "red", "grey) respectively)}

\item{inc.equal}{logical whethere points on the line should be counted as above (dafault=TRUE)}

\item{labels}{label for the markers (default=c("above", "below"))}

\item{type}{type of plot for data (options include point (dafault), line, stepwise)}
}
\description{
This function catagorises data points into above or below a cut-off plane
}
\examples{
data(beavers)
df.plane = beaver1
df.data = beaver2
df.plane$temp <- df.plane$temp +0.5
bio_bires(x="time", y="temp", df.data, x.plane="time", y.plane="temp", df.plane)
}
\keyword{plane,}
\keyword{scatter}
