% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bio_boxplots.R
\name{bio_boxplots}
\alias{bio_boxplots}
\title{Comparison Boxplots}
\usage{
bio_boxplots(data, x, y, p.cutoff = 0.5, stars = FALSE,
  method = "t.test", star.vals = NULL)
}
\arguments{
\item{data}{}

\item{x, y}{x and y variable names for drawing.}

\item{p.cutoff}{plot p-value if above p.cutoff threshold. To include all comparisons set as NULL.}

\item{stars}{Logical. Whether significance shown as numeric or stars}

\item{method}{a character string indicating which method to be used for comparing means.c("t.test", "wilcox.test")}

\item{star.vals}{a list of arguments to pass to the function symnum for symbolic number coding of p-values. 
For example, the dafault is symnum.args <- list(cutpoints = c(0, 0.0001, 0.001, 0.01, 0.05, 1), symbols = c('****', '***', '**', '*', 'ns')).
In other words, we use the following convention for symbols indicating statistical significance:
ns: p > 0.05; *: p <= 0.05; **: p <= 0.01; ***: p <= 0.001; ****: p <= 0.0001}
}
\description{
Creates boxplots showing the significance between groups. Adapted from kassambara/ggpubr.
}
\examples{
bio_boxplots(iris, x="Species", y= "Sepal.Width", p.cutoff = 0.0001)
bio_boxplots(iris, x="Species", y= "Sepal.Width", NULL, stars=TRUE)
}
