% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bio_expcomp.R
\name{bio_fc_heatmap}
\alias{bio_fc_heatmap}
\title{Fold change and pvalues between groups}
\usage{
bio_fc_heatmap(exp, var, prefix = "", stars = FALSE, overlay = TRUE,
  logp = FALSE)
}
\arguments{
\item{exp}{Expression Data}

\item{var}{Vector classing samples by variables}

\item{prefix}{Prefix to Heatmap titles}

\item{stars}{whether pvales should be written as numeric or start (default=FALSE)}

\item{overlay}{pvalues on fold change Heatmap or besidde (default = TRUE)}

\item{logp}{Whether or not to log the pvalues}

\item{...}{Other parameters to pass to Complex Heatmap}
}
\value{
A list containing the mean expression for each group, the fold change between groups, the pvalues comparing the expression from different groups, a heatmap with mean expression, a heatmap containing the fold change and p-values.
}
\description{
This function exports a complex heatmap looking at the expression of different modules which can be customised
}
\keyword{change,}
\keyword{expression,}
\keyword{fold}
\keyword{heatmap,}
\keyword{pvalues}
