% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bio_expcomp.R
\name{bio_fold.change}
\alias{bio_fold.change}
\title{Fold Change between Groups}
\usage{
bio_fold.change(exp, var)
}
\arguments{
\item{exp}{Expression Data}

\item{var}{Vector classing samples by variables}
}
\value{
data frame with mean expression levels
}
\description{
This function calculates fold change in expression between sample groups
}
\keyword{change,}
\keyword{expression,}
\keyword{fold}
\keyword{heatmap,}
\keyword{pvalues}
