% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bio_frequency.R
\name{bio_frequency}
\alias{bio_frequency}
\title{Function to generate a frequency table}
\usage{
bio_frequency(data, columns, freq.percent = "both", include.na = TRUE,
  remove.vars = NULL)
}
\arguments{
\item{data}{A data frame containing columns to be counted}

\item{columns}{Column names or indices to be counted in data}

\item{freq.percent}{Whether the table should include frequency counts, percentages or both (options = c("freq", "percent", "both")). Default="both"}

\item{include.na}{Include NA values (options are TRUE/FALSE, default=TRUE)}

\item{remove.vars}{Character vector of variables not to be included in the counts (e.g. remove.vars = c("") remove blanks from the count)}
}
\description{
This function generates a frequency table from factor or character vector columns in a data frame
}
\examples{
bio_frequency(mtcars, columns=c("vs", "am"))
}
\keyword{counts,}
\keyword{frequency,}
\keyword{table}
