% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bio_geneid.R
\name{bio_geneid}
\alias{bio_geneid}
\title{Function to switch Gene Ids}
\usage{
bio_geneid(IDs, IDFrom = "ensembl_transcript_id", IDTo = "hgnc_symbol",
  mart = "ensembl", dataset = "hsapiens_gene_ensembl",
  attributes = c("chromosome_name", "start_position", "end_position"))
}
\arguments{
\item{IDs}{list of the Ids you want to convert}

\item{IDFrom}{What format these IDs are in (default Ensembl)}

\item{IDTo}{What format you want the IDs converted to (default gene names)}

\item{mart}{The biomart to use. Typically, for humans you will want ensembl (default). Alternatives can be found at listEnsembl()}

\item{dataset}{you want to use. To see the different datasets available within a biomaRt you can e.g. do: mart = useEnsembl('ENSEMBL_MART_ENSEMBL'), followed by listDatasets(mart).}

\item{attributes}{list of variables you want output}
}
\description{
This function switches gene (or snp) ids using biomart
}
\examples{
IDs = c("TNF", "A1BG", "FOX3")
bio_geneid(IDs, IDFrom='hgnc_symbol', IDTo = 'ensembl_transcript_id')
}
\keyword{ids}
