% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bio_expcomp.R
\name{bio_mean.by.var}
\alias{bio_mean.by.var}
\title{Mean Expression across Groups}
\usage{
bio_mean.by.var(exp, var)
}
\arguments{
\item{exp}{Expression Data}

\item{var}{Vector classing samples by variables}
}
\value{
data frame with mean expression levels
}
\description{
This function calculates the mean expression (z-score) of sample types
}
\keyword{change,}
\keyword{expression,}
\keyword{fold}
\keyword{heatmap,}
\keyword{pvalues}
