% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bio_modscores.R
\name{bio_modscores}
\alias{bio_modscores}
\title{Calculate the module score for expression data}
\usage{
bio_modscores(exp, module.list)
}
\arguments{
\item{exp}{Expression Data}

\item{module.list}{List of genes in each module}
}
\value{
data frame with expression levels for modules
}
\description{
This function calculates the mean expression (z-score) of sample types
}
\keyword{expression}
\keyword{module,}
\keyword{svd,}
