% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bio_sankey.R
\name{bio_sankey}
\alias{bio_sankey}
\title{Longitudinal Sankey}
\usage{
bio_sankey(samp.order)
}
\arguments{
\item{samp.orders}{A list of named vectors for sample order at each timepoint. Vector names must correspond to matchable
ids.}
}
\description{
Creates a sankey plot with heatmaps showing how individuals progress over time.
}
\examples{
library(survival)
data(kidney)
df <- kidney
library(lme4)
data(sleepstudy)
data = sleepstudy
data = data[data$Days !=1 | data$Subject != "308", ]
time.col="Days"
exp.cols = "Reaction"
sub.col = "Subject"

row.order=list()
for(i in unique(data[, time.col])){
 hm = Heatmap(data[data[, time.col] == i, exp.cols], cluster_rows=cluster)
 row.order[[paste('timepoint', as.character(i))]] = setNames(unlist(row_order(hm)), data[data[, time.col] == i, sub.col])
}

bio_sankey(samp.order=row.order)
}
