% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bio_treatmentGroups.R
\name{bio_treatmentGroups}
\alias{bio_treatmentGroups}
\title{Function to compare two treatment groups}
\usage{
bio_treatmentGroups(df, group.col, y.col = "y", x.col = "x",
  id.col = "id", main = "", cols = NULL, p.col = "maroon")
}
\arguments{
\item{df}{Data frame containing the data for both groups}

\item{group.col}{Column name which corresponding to the group of values in df}

\item{y.col}{Column name corresponding to the y-axis values in df}

\item{x.col}{Column name corresponding to the x-axis values in df}

\item{id.col}{Column name which corresponds to subject ID}

\item{main}{Title of pot}

\item{cols}{Character vector for colours of lines}

\item{p.col}{Colour of p-value text}
}
\description{
This function plots a line graph comparing two treatment groups over time
}
\keyword{line}
