% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bio_volcano.R
\name{bio_volcano}
\alias{bio_volcano}
\title{Function to generate a volcano plot}
\usage{
bio_volcano(toptable, fc.col = "log2FC", padj.col = NULL,
  padj.method = "fdr", padj.cutoff = 0.05, fc.cutoff = 1,
  marker.colour = c("grey60", "olivedrab", "salmon", "darkturquoise"),
  label.p.cutoff = NULL, label.row.indices = NULL,
  label.colour = "black", legend.labs = NULL, add.lines = TRUE,
  line.colour = "grey14", main = "Volcano Plot", xlims = NULL,
  ylims = NULL)
}
\arguments{
\item{toptable}{A data frame containing p value and fold change columns for parameters compared across multiple groups. The p value column should be named "pvalue".}

\item{fc.col}{The column name which stores the fold change. Should be in the log2 format (default="log2FC")}

\item{padj.col}{The column  which contains adjusted p-values. If NULL adjusted pvalues will be calculated}

\item{padj.method}{correction method. Options include: c("holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none"). Default="fdr}

\item{padj.cutoff}{The cutoff for adjusted pvalues. This adds a horizontal line of significance (default=NULL)}

\item{fc.cutoff}{The log2(fold change) significance cut-off (default=1)}

\item{marker.colour}{Character vector of four colours to map to the volcano plot. In the order non-significanct, fold-change significant, pvalue significant, significant in fold-change and pvalues (default=c("grey60", "olivedrab", "salmon", "darkturquoise"))}

\item{label.p.cutoff}{The cutoff for adjusted pvalues for labelling (default=NULL). Not recommended if many significant rows.}

\item{label.row.indices}{Indices of rows to be labelled (default=NULL)}

\item{label.colour}{Colour of labels (default="black")}

\item{legend.labs}{A character vector for theThe legend label names (default=c("Not Significant", "FC>fc.cutoff", "Padj<padj.cutoff", "FC>fc.cutoff& Padj<padj.cutoff"))}

\item{add.lines}{Whether to add dashed lines at fc.cutoff and padj.cutoff (default=TRUE)}

\item{line.colour}{The color of dashed significance lines (default="grey14")}

\item{main}{Plot title}

\item{xlims, ylims}{The plot limits}
}
\description{
This function generates a volcano plot from a top table using ggplot.
}
\examples{
bio_volcano(toptable)
}
\keyword{volcano}
