% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bio_popcomp.R
\docType{data}
\name{pop}
\alias{pop}
\title{Function to compare multiple populations}
\format{
An object of class \code{character} of length 1.
}
\usage{
pop
}
\arguments{
\item{data}{A data frame containing x, y, and population data}

\item{xlab}{Column name or index of the x variable}

\item{ylab}{Column name or index of the y variable}

\item{poplab}{Column name or index of the population variable}
}
\description{
This function generates a volcano plot from a top table using ggplot.
}
\examples{
bio_volcano(toptable)
}
\keyword{volcano}
