% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bio_gene_locations.R
\name{bio_gene_locations}
\alias{bio_gene_locations}
\title{Gene Locations}
\usage{
bio_gene_locations(
  chromosome = 6,
  xrange = c(33500000, 3.4e+07),
  subset_genes = c(),
  ggtheme = theme_null(),
  output_plots = TRUE
)
}
\arguments{
\item{chromosome}{Which chromosome are you looking at}

\item{xrange}{chromosome position}

\item{ggtheme}{ggplot theme to give plots.}

\item{output_plots}{Logical whether to output plots}

\item{subset_gene}{List of genes to plot/subset to within xrange}
}
\description{
This function produces ggplot and plotly plots showing the genes in a 
genomic range
}
\examples{
bio_gene_locations()
}
