% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bio_mods.R
\name{bio_mods}
\alias{bio_mods}
\title{Function to investigate expression for custom modules}
\usage{
bio_mods(exp, mod.list, meta, mean.var, cluster.rows = TRUE,
  cols = c("red", "blue", "green3"), main = "", show.names = TRUE)
}
\arguments{
\item{exp}{data frame containing the expression data}

\item{mod.list}{A list of modules. Each element contains the list of genes for a modules. The gene names must match the rownames in the exp dataframe.}

\item{meta}{Dataframe where each column contains an annotation/tract for samples in the heatmap. The order of samples in meta must match that of exp.}

\item{mean.var}{Character defining the meta column to average (mean) over.}

\item{cluster.rows}{The method to use for clustering of rows}

\item{cols}{Chacter vector, or named vector to fix the order, defining the colours of each mean.var group.}

\item{main}{Title of heatmap}

\item{show.names}{Show row names. Logical.}
}
\description{
This function exports a complex heatmap looking at the expression of different modules which can be customised
}
\keyword{heatmap,}
\keyword{module}
